Added race updates via time series topic.
The race topic is 'race/updates/fast' and contains a time series datatype of JSON.
The JSON is formatted like this: "[{"pos":0.0},{"pos":0.0}]"
Basically an array containing all cars in the race and relevant data of those cars.
At the moment it'll just be the position on the track in the range of 0.0 to 1.0.